Move dashboard constants to common and update imports
