+++ src/pages/Live/components/NotesBlock.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from '../../common/SensorDashboard.module.css';
 import idealRangeConfig from '../../../idealRangeConfig.js';
-import {bandMap, knownFields} from '../../Dashboard/components/dashboard.constants.js';
+import {bandMap, knownFields} from '../../common/dashboard.constants.js';
 
 function NotesBlock({ mergedDevices = {} }) {
   const metaFields = new Set(['timestamp', 'deviceId', 'compositeId', 'layer']);
